+++ tom_d4rt_generator/lib/testing.dart
@@ -26,4 +26,5 @@ library;
 export 'src/testing/d4rt_test_result.dart';
 export 'src/testing/d4rt_tester.dart';
 export 'src/testing/example_freshness.dart';
+export 'src/testing/package_resolution.dart';
 export 'src/testing/issue_test_helper.dart';